Favor new connections over old connections
When multiple listening clients connect with the same name, connecting to that
name now connects to the most recent client, rather than the oldest.

When a listening client reconnects without properly shutting down the old
connection, this results in multiple connections with the same name. Before
this commit, this would prevent connecting the the client, which is now
possible again.