Use process_factory for creating SSHServerProcess objects
Previously, the session_requested callback would create a SSHServerProcess
object directly, but its constructor is not actually public API (so was changed
in a breaking way in asyncssh 2.9.0). This sidesteps the issue by letting
asyncssh create the object instead.

Unfortunately, the only way to do this is to use a global process_factory,
rather than some SSHServer callback. To fix that, this adds a small global
callback that figures out the server object associated with the requested
process, and forwards handling to that server.