Let SSHServerConnection handle more of key authentication
This passes the authorized_keys file to the connection object to handle on its
own, and relies on the auth_completed() callback to process the resulting
key options (which are saved by the connection code).

This allows removing the `public_key_auth_supported()` method, whose default
implementation is now fine.

The `validate_public_key()` option is still kept, since it will be called when
a key is not found in the key file and still allows logging rejected keys
(accepted keys are no longer logged after this change).

This also results in a small bugfix: Previously, the username would
(incorrectly) be used in place of the `peer_host` when calling
authorized_keys.validate(), now the actual `peer_host` (which is probably
a reverse lookup on the IP address done by SSHServerConnection) is used as
should be. This probably affects keys that had a "from" option set to restrict
usage to certain clients.